Extract test classes to separeted files.
